Add "data" package.
This commit adds all files from the "data" package in the original version of
this project. I have made minor changes to the files to make them compatible
with this version of the project and to fix various errors, but I have not
tested them. They should be thoroughly reviewed, improved, and tested at a
later date.