Fix a lexicon parser crash.
Fixes #64. The parser was crashing due to an empty "tr" element. The remaining
oddities in the text mirror the entry displayed on Perseus, with one
exception: the "key" value, which we use to display the entry heading, is
incorrect. We need to use the "orth" value instead. I have created issue #65
to track this bug.