Replace Android Icons with Material Design icons
Replaces all icons except for the app icon with images from Google's
Material Design Icons set. This commit also adds numerous icon files in
alternate colors that are not currently used in the app. This is being
done for the sake of convenience in the expectation that we may need
some of these icons for future UI updates; I'll clean up any remaining
unused icons once the Material Design update work is finished.

Closes #108 and addresses #89.