Add typeface preference
Adds a preference allowing the user to choose between the Noto Serif and
Roboto typefaces on devices running SDK 21 or higher. Older devices
cannot use Roboto due to lack of support for polytonic Greek characters.
This commit also includes various improvements to the design of the
SettingsActivity and GreekTextView classes.

Closes #107.