Fix a crash when rotating screen
Fixes #142. It looks like getActivity() was sometimes returning null
on screen rotation because the parent activity had been destroyed and
not yet recreated. We now save a reference to the activity in
onAttach(). Unfortunately I can't reproduce this bug, so I can't verify
with certainty that this will fix it, but I think this will do the
trick.