Throw exception on missing lexicon entry
Previously, if a lexicon entry could not be found in
LexiconDetailFragment#onCreate(), the resulting exception was caught and
a message was written to the Android log. mLexiconEntry remained null,
causing a later exception. A NullPointerException will now be thrown in
this scenario, so we will hopefully receive bug reports from users when
errors happen here going forward.

This is obviously a less than optimal way to deal with this bug.
Long-term, we need to add automated testing to find and fix lexicon
parsing problems. It would also be nice to catch and gracefully handle
errors of this sort, propmting the user to submit a bug report, rather
than crashing.

See bug #148.